Update PTP enable state correctly in SAI
Summary: Current update of PTP enable logic works if its invoked after the port is added. But it doesn't work if the PTP is enabled from start and port(s) are added later. Cache the PTP state in switchManager and use it when new port is added. Both scenarios of enable/disable PTP are covered.

Differential Revision: D34334890

fbshipit-source-id: 538da4cc597e6a62ecfa64fab42e4999b545ffd7