Fix to send max 32 byte key to driver for creating SA
Summary:
We are expected to send 32 bytes key to the driver. From the MKA, this 32 byte key comes as a 64 character string. We need to convert this 64 char string to 32 byte array by treating 2 character as 1 byte but our code was treating 1 character as 1 byte so this got converted to 64 byte array instead of 32 byte array. When this 64 byte array goes to the driver then it causes some issue like memory corruption while creating new SA. Due to this the subsequent call to install ACL fails with bogus error NOT_IMPLEMENTED.

The driver expects application (us) to verify the key is maximum 32 byte array which we were doing wrong and we were passing the 64 byte to it. I have put the fix to convert this string to array by converting 2 hex characters to 1 byte. This fix works in my standalone system where I was able to recreate the exact same issue. I will also ask vendor to throw error in this case rather than do memory corruption silently.

Reviewed By: jasmeetbagga

Differential Revision: D34385206

fbshipit-source-id: 42b91b83d16be9df256bd869f4479f0e61a872df