Fix thrifty migration traverse logic
Summary:
In ThriftyNodeMapT::migrateToThrifty my logic for traversing down the dynamic was reversed. Need to traverse if kThriftySchemaUpToDate is false but I had the opposite. Only hit this problem with vlan because its the first time we have a node map of node maps and the deeper levels need the migration.
Introduced a helper to make this easier to read and prevent further errors

Reviewed By: jasmeetbagga

Differential Revision: D34356115

fbshipit-source-id: b8b094cbc15651b578dcb10d9625d4c18496bbde