Introduce map storing QPH multiple nexthops and handle in addRoute
Summary:
This is towards supporting QPH monitoring for routes with multiple next hops. A map of <prefix, set<neighbors> is introduced to store prefixes and their next hops containing classIds. The high level idea is -
1. In `addRouteAndFindClassID`, if a route contains two next hops with classIDs, add it into the map, else remove the entry from the map if no longer a multi-nexthop route. (This diff)
2. In the case of neighbor gets unresolved, remove entry from the map and check if it's no longer a multi-nexthop route. (D33662783)
3. In the case of neighbor gets resolved (and neighbor has classId), we need to check -
a. If already a multi-nexthop route, add neighbor entry into the map.
b. b. If both route and the new neighbor both has classID associated, add them into the multi-nexthop map.(D33662812).

Unit tests are added in D33690820. Will add ODS counter to publish the number of multi-nexthop routes when there's a change in the map size.

Note that this is only to cover routes will multiple next hops with different classId. TODO: add check for route with QPH neighbors and non-QPH neighbors.

Reviewed By: shri-khare, jasmeetbagga

Differential Revision: D33662070

fbshipit-source-id: e110f1e3d8d3fc0e2808bd81fd20a8aed1cb330a