rackmon: Add support for register maps and interpretation
Summary:
Add capability to read a register map JSON configuration
which provides information of:
1. What registers to monitor and read (address, begin)
2. How big the registers are
3. How to interpret the contents of each register.
Other than this, this provides a mechanism for there to
be multiple "register maps" depending on the address of
the modbus device -- A register map (A JSON File) specifies
the range of modbus addresses for which this register map
applies. So, for example, this allows us to apply unique
register maps to PSU devices, BBU devices, ORv3 PSU devices etc.

Reviewed By: tao-ren

Differential Revision: D34095929

fbshipit-source-id: 6e05634df9bb4d4fbe0d964b3fd481e62280de55