Derive SffFieldInfo from QsfpFieldInfo
Summary:
as titled. A base class QsfpFieldInfo was added in D34676577. In this diff, SffFieldInfo is changed to derive off that.

Also made SffPages as enum class instead of enum so that it doesn't interfere with the same enums in CmisModule. Also assigned values to the pages since I'll need the value of these pages later on. The lower page is assigned -1, and the rest of the pages are assigned their corresponding positive integer.

Differential Revision: D34676582

fbshipit-source-id: 043426d2a149fcbfb3ee4b9fdc20a87761dbbda7