Derive CmisFieldInfo from QsfpFieldInfo
Summary:
as titled. A base class QsfpFieldInfo was added in D34676577. In this diff, CmisFieldInfo is changed to derive off that.
Also made CmisPages as enum class instead of enum so that it doesn't interfere with the same enums in SffModule. Also assigned values to the pages since I'll need the value of these pages later on. The lower page is assigned -1, and the rest of the pages are assigned their corresponding positive integer.

Differential Revision: D34677050

fbshipit-source-id: 424d75c8a1bf1ee5d066872f6ad5b582f954624c