Fix destruct + cancellation race is FsdbStreamClient and derivatives
Summary:
Upon cancel we schedule a cancellation of service loop.
This will happen in streamEvb. However, we do not synchronously wait
for this cancellation to happen. This means that if b/w the time of
signalling cancellation and actual cancel, if we destroy the
object we can get a dangling memory access

Differential Revision: D34589589

fbshipit-source-id: 186fa27e255bf1674329a80b33e546c18706bdf5