Periodically update prbs stats
Summary:
Our thrift interface for getting prbs stats returns back a PrbsStats object which has fields like numLossOfLock. The intended flow is

1. Thrift API to set Prbs
2. wait some time
3. Thrift API to get PrbsStats

To collect information such as numLossOfLock or maxBer we need to periodically read prbs stats from the module. Therefore in this diff, i add a updatePrbsStats which is called in refreshTransceivers logic where all the other data is read from the transceivers. The prbs stats are only read when prbs is enabled. Once these stats are read from the respective implementations (SffModule/CmisModule), the common code in QsfpModule will update the fields like numLossOfLock or maxBer. The logic for this in updatePrbsStatEntry lambda will be added in the next diff

Reviewed By: rajank7

Differential Revision: D33902140

fbshipit-source-id: a9b85c78338867d26167fed3a1319201d5d78d9c