reset sample packet for th3
Summary:
As titled, reset sample packet for TH3.

In TH3, SaiPort object is marked not live (releasePorts) in sai store as part of sai port manager destructor. After marking not live, the port handle destructor is called which tries to reset the OID for samplepacket on this sai object which causes a core.

Reset the sample packet OID on a port for TH3 explicitly.

Reviewed By: shri-khare, zechengh09

Differential Revision: D34796175

fbshipit-source-id: 2b336ef4e6603b29dce9f8b7c6e1b9db2b8b77ef