LookupClassRouteUpdater: block MAC for non-MH-NIC
Summary:
[this diff is originally drafted by shri-khare]

Adding support to cache non-MH-NIC routes once included in the block list.

For normal routes that are not pointing to MH-NIC (minus some other routes), they are currently not cached by `subnet2VlanCache`, hence we're not able to assign drop classID to them.

To address this issue, once a Mac address is included in the block list, we'll iterate through the neighbor table to find all IPs that are resolved to this MAC, and cache the routes to the `subnet2VlanCache`.

This diff introduces the outline of `processMacAddrsToBlockUpdates`. Will follow add and remove logic in the next two diffs.

Reviewed By: shri-khare

Differential Revision: D34404166

fbshipit-source-id: 610105324c7126de80ee09011e131d8235c796c1