use libcurl in getdeps
Summary:
When using getdeps inside of a container, Python's urllib isn't able to download from dewey lfs (see this post for details https://fb.workplace.com/groups/systemd.and.friends/permalink/2747692278870647/).

This allows for getdeps to use `libcurl` to fetch dependencies, which allows for a getdeps build to work inside the container environment.

Differential Revision: D34696330

fbshipit-source-id: 1a25608badd02f56c5d4abd3d19b205107d7a383