get route metadata from SDK before updating next hop
Summary:
Consider a scenario
1) link goes down
2) route is reset to CPU
3) before link down handling is processed and neighbor/next hop is removed, agent is shutdown
4) agent warms boot
5) route object is reloaded from store with a metadata of 1 or 2
6) since neighbors were not resolved/next hop were not removed, the route points back to next hop
7) at this point metadata is reset to 0 by SDK
8) update metadata finds expected metadata as 2 and actual metadata as 0
9) difference triggers write to SDK with metadata of 2, this leads crash

metadata changes as next hop is written to route. if route was pointed 2 and now resolves to next hop, and if this happens over warm boot, expected metadata would be 2 and programmed metadata would be 0 in SDK if programmed metadata is queried after route is pointed to next hop.

Reviewed By: zechengh09

Differential Revision: D34728111

fbshipit-source-id: cbdb96f4f14909e41311766eb88d50ada6582b0c