NDP solicit neighbor test
Summary:
**Issue**: When we send NDP neighbor solicitation(NS) packet to a multicast IPv6
address(ff02:0:0:0:0:1:ff00::/104 appended with last 28 bits of neighbors unicast IPv6 address), we receive it back on CPU port. This is an issue because, consider an RSW for which we have unresolved servers connected. And we keep receiving packets/traffic destined to those servers. This will result in NDP originating NS packets, and those will flood the CPU queue, impacting other control packets.

**Root Cause**: We have installed NDP solicitation ACL after the ACL that catches the multicast packets from CPU and only forward it, but NOT copy to CPU. This will be more clearer in the Test plan.

**Fix**: Add the ACLs to config in right order.

Reviewed By: jasmeetbagga

Differential Revision: D34782575

fbshipit-source-id: 61ba0bd2da529de6f250dd93439d9c228a972b94