add queue config and acl to sflow test
Summary:
As titled, add queue config and acl to sflow test.

For Tajo, there is a requirement of atleast one ACL in the hardware for the traps to work. if not, there is a default ACL installed which gets priority over the trap.

Queue config is necessary to pull the queue stats. Queue stats are not fetched otherwise.

Reviewed By: jasmeetbagga

Differential Revision: D33968374

fbshipit-source-id: f4894e1c9988a898a7b5402ec8a6ac51da45ae6e