rackmon2: Remove support for deprecated string/print data
Summary:
After "value" data was added which is a structured interpretation
of the value stored in the registers, it formally removed
any value of having "print format". But since we did not have
CLI properly or usably print the "value" data, it was held
temoprarily.

Now that CLI implements a proper parsed/formatted display of
the value data, we can formally remove this functionality.

For Thrift all this was dead code anyway.

Imported from: D34598738

Reviewed By: jasmeetbagga

Differential Revision: D34707794

fbshipit-source-id: 98d11b873a8158553fc6d22460361fd3ad3821a7