rackmon2: Move external events into PollThread
Summary:
This adds a functionality into PollThread to allow
for an external entity to trigger events and not
just rely on polling. This enables us from building
better mocks instead of having to create a MockPollThread
for each and every polling thread instance. Plus the feature
seems useful especially if I want to introduce a "force scan now"
feature.

Reviewed By: jasmeetbagga

Differential Revision: D34978241

fbshipit-source-id: ebb4c18e2fc50c8ed0a9f89a2b22b1772d1766de