Remove cmisModuleReady field from old state machine
Summary:
It doesn't look like we never used `cmisModuleReady` in the legacy state machine anywhere besides
CmisModule::updateQsfpData().
But in this function, it always reads from the qsfp register anyway and then uses that to decide whether to
cache page14 and VDM related pages.
Therefore, we don't actually need to store this state, especially we're going to deprecate the old state machine anyway.

Differential Revision: D34975886

fbshipit-source-id: 7f40d1c89b6bc5a35f73d06d160916cfb6d2538d