rackmon2: Clean up req/resp messages upon command failure
Summary:
Modbus could have encoded/decoded the message or in other words
changed the length of the contents.
Req: We are guaranteed that encode is either non-stateful (Can be
repeated with the same initial len or wont change raw[] position if
some bytes are present (The only use of this is WriteMultiple)).
Resp: The only thing we want from this is the length of the
response expected, so in this case, we are good.

Imported from: D35062413

Reviewed By: jasmeetbagga

Differential Revision: D35062791

fbshipit-source-id: a08059aa26868276b0462cdf3f195435107c39fe