Fix flakiness of verifyWredThreshold
Summary:
Currently, verifyWredThreshold is waiting for expected number of packets to be seen as out packets before continuing
with the tests. However, at this point, the wred drop counters might not be updated. To avoid this case, make sure to
check for drop+out stats to reach expected value before continuing with tests.

Reviewed By: jasmeetbagga

Differential Revision: D35161239

fbshipit-source-id: ca783ec27bc632899e504968126c4a8758875827