don't fail if no archived logs are found
Summary: We've seen a few cases where we fail to fetch snapshots. One reason for this is that `ls FILE` will fail if FILE doesn't exist. On a newly provisioned switch, we won't have any archived logs yet, so the existing LS command would fail. As such, I'm changing the code to instead `find` the archive folder and grep for the archive logs. Now this should just return an empty list of log files if no such files are present.

Differential Revision: D35086342

fbshipit-source-id: 0669bae6e0d8d85a45866ced2d709cce874e2f2f