add LOG_THRIFT_CALL_MAYBE_INLINE
Summary:
This adds a new macro for logging if the code could be
executed in either a thrift server or inline via a cli. This is the
case for fboss_updater, where we support both as a service and as a
cli tool that directly instantiates a Handler() object.

In order to still get some info if a user ssh's in and calls the
fboss_updater api directly, we fall back to reading from the SUDO_USER
environment variable. This is obviously not 100% reliable, but since
this is only used for logging and not authentication it should be
better than always just outputing "unknown".

Reviewed By: echistyakov

Differential Revision: D35189313

fbshipit-source-id: c7d21c2593da68109f05dd0f950d9a36f5423ce4