Introduce new state machine unit test
Summary:
After D34980472, we still have a bunch of unit tests for the legacy state machine,
instead of just deprecate all these tests, I decided to introduce a new set of unit tests for the new state machine.

Hopefully we can use these tests to catch the core state machine logics, regardless of the real hardware programming,
which should be taken care in our qsfp_test and link_test already.

In this diff we'll verify the logic of the first events DETECT_TRANSCEIVER

Differential Revision: D35068649

fbshipit-source-id: 7e15883f610a21fb808b2a87e0bae85cdfa88858