resetProgrammedIphyPortToPortInfo() after transceiver reset
Summary:
When I tried to introduce a new unit test to mimic programming iphy port failure, I noticed we didn't clear the
cached ProgrammedIphyPortToPortInfo after we removed the transceiver.

To fix that, call resetProgrammedIphyPortToPortInfo() in resetProgrammingAttributes() so that every time we reset
a transceiver, the cached info will be clear too.

Differential Revision: D35068645

fbshipit-source-id: 3eb1ca38561a324ed878b8ccc798b336851bc8c5