Onboard Scheduler API to new logging scheme
Summary:
As titled. We see the limitation here that uint64 has a collision with sai_object_id_t, since they're both the same type (and we use type index in the new scheme).

Functionally this is the same, since attribute value is a union. But that's mostly affecting readability of the generated logs.

Will think of some other ways to take care of this.

Differential Revision: D35159869

fbshipit-source-id: 4c286e86fa949758a66242e030a05e072c316b0f