Avoid transceiver reprogramming if CMIS reaches the max remediation times
Summary:
For CMIS module, current remediation logic has a limit of hard reset times.
And once we reached the limit, the CmisModule::remediateFlakyTransceiver() is actually a no-op
therefore, we shouldn't treat it as a successful remediation and def should avoid reprogramming transceiver again.

Use bool as a return for remediateFlakyTransceiver()

Differential Revision: D35323217

fbshipit-source-id: 25d50186070ca157da051a2b1dee27e1b6a567a8