Verify remove transceiver logic in new state machine
Summary:
This diff introduced a new unit test to verify the `TransceiverStateMachineEvent::REMOVE_TRANSCEIVER` logic from the new state machine.

I also decided to remove the state machine guard of this event when the previous state is `DISCOVERED` and `PRESENT`. Because the state machine hasn't started the actually port programming yet, there's no reason for us to ask for this all ports down check.

Differential Revision: D35108162

fbshipit-source-id: 5ce7fe10fc5d6713561e34952724b3ac726aca2b