getdeps boilerplate setup
Summary:
Stack context:

To get a build setup for edencommon I picked an eden library to move over
to edencommon.

I picked process name cache. I would like to have this library in watchman
so that I can log spawning process command lines as well as client command
lines.

Diff:
This diff sets up the boiler plate for cmake and getdeps so that we can build
edencommon in oss

Reviewed By: chadaustin

Differential Revision: D34218025

fbshipit-source-id: 0c75e9ab2d5a7fecce09405b27cb5d1eb49c938d