Introduce WITH_RETRIES and ASSERT_EVENTUALLY
Summary:
The motivation here is to replace usage of checkWithRetry in tests. checkWithRetry doesn't really get us good logs as it just throws an exception and we lose a lot of info about what actually failed. Specifically when we have multiple checks inside a single checkWithRetry, or when we have multiple calls to checkWithRetry, its a bit hard to pinpoint which specific check failed.

The idea is to try to leverage gtest utils as much as possible to get nice logs and actual line numbers on failures. See comment on how this would be used

Reviewed By: jasmeetbagga

Differential Revision: D35306376

fbshipit-source-id: 997dbf42f572020f442fea2b4c6eca28022a760f