Support hard reset present transceiver if can't detect management interface
Summary:
There's a ad-hoc hard reset transceiver logic in current WedgeManager::updateTransceiverMap(), where if the transceiver is present but not able to detect the management interface, it will trigger a hard reset.

However, the current logic is using WedgeManager::ports_, which is no longer needed for the new state machine.
Use TransceiverManager::areAllPortsDown() directly.

Differential Revision: D35379225

fbshipit-source-id: 21203b5841f9afdf5b30b065f8b0d064c2e7ea71