Introducing FsdbPublisherMetdataTracker
Summary:
Main idea behind this class is to track metadata for
each of the publishers that have open streams. We will then use
this to report metadata from publishers to subscribers. We will
also use this to convey publisher disconnect to subscribers
when last of the publisher streams goes away

Differential Revision: D35384451

fbshipit-source-id: 59573c09494beb30aca1ce75b548a418e757ae48