Added skeleton for Prbs link test
Summary:
This stack will add a link test that will test the prbs functionality on all supported components - ASIC, GB, OPTICS. Currently, I'll be adding support for optics only, but trying to design the test so that it can easily be extended to other components later.

In this diff, I added a skeleton for the prbs test base class and then defined a transceiver prbs test that derives off it. We'll have multiple test cases per optic type , per component and per polynomial. The known bad test config can mark the test known bad on platforms that don't support prbs.

Differential Revision: D35351855

fbshipit-source-id: 4c2c4cf02668bc7fabb71c77976acfe8c106e64c