sort timestamps before filtering
Summary: the timestamp filtering logic assumes the timestamps are sorted, as it goes in order through the list and stops when it gets past the last log file that could contain the timestamp. Previously we used `ls` to query the log file names, but since we switched to `find`, it broke since `find` doesn't sort its output. (I definitely should have just sorted the list in the first place instead of having this requirement be implicit). This diff sorts the timestamp list before filtering to fix the issue

Differential Revision: D35444082

fbshipit-source-id: 58922678ee8777739442901911bbf4e2a6ad12b6