Filter out transceivers with the given media type and which support PRBS for the test
Summary: as titled. Filter out relevant transceivers for the transceiver prbs test. Currently only supports testing PRBS on the transceiver line side. For checking if PRBS is supported or not, I use the getSupportedPrbsPolynomials thrift API which currently only qsfp_service supports but will also be added to wedge_agent in future

Differential Revision: D35352182

fbshipit-source-id: ec046f3511768ca15ffef88f8264ad5216939b15