Pivot from tracking metadata per pulisherId to tracking for publisher roots
Summary:
In FSDB we want to track metadata for each of the pusblisher
roots (viz. agent, qsfp_service, bgp). Previously I used publisherId to
track this. It was convenient that publisher IDs happened match publisher
root path names. But this is not a hard requirement. Further we would
like to server this metadata to subscribers as well. But the publisher
info is lost when serving subsciptions (viz. if mulitple publishers
published under a root, which publisher's metadata should we use?).
So do the right thing and track metadata for publisher roots itself.

Differential Revision: D35453564

fbshipit-source-id: e085b21d739a0924ae7eb791c411ec75f965a674