Only fetch cmisStateChanged bit when transceiver is present
Summary:
As title.
If we don't check whether the transceiver is present, getModuleStatus() will throw exception:
```
Qsfp is either not present or the data is not read
```
when trying to read the register on an absent transceiver.

This will cause the refresh() function throw exception and not finish the rest of refresh logic.

Reviewed By: birdsoup

Differential Revision: D35482498

fbshipit-source-id: 4a27655c8faede0136dcb703320b43e2814a8a76