sandia 400G port with 4 lanes
Summary:
As titled, generate sandia platform mapping with 4 lanes for 400G. This is because sandia has 100G serdes.

NOTE: Currently, sandia uses the existing port profiles and sets the number of lanes to be 4 to generate the config. The right way is to add new port profiles to switch config to explicitly state the serdes and number of lanes.

Reviewed By: jasmeetbagga

Differential Revision: D35581770

fbshipit-source-id: 5b6df697f4239e1b6a3e2cfacb764654fe4bebfe