Move responsibility of synchronizing FsdbOperTreeMetadataTracker to clients
Summary:
For pruning subscribers when all publishers for it go away,
we need to look at a FsdbOperTreeMetadataTracker post a publisher
unregister. Then we need to check if this was the last publisher for
a publisher root and if so notify corresponding subscribers. All this
needs to be done while ensuring that FsdbOperTreeMetadataTracker does
not change (viz. if a new publisher registers b/w call to unregister
and us checking, we could miss a all publishers gone for this root
event). So move the responsibility of synchronizing FsdbOperTreeMetadataTracker
to clients, where we can control how long to hold locks for.

Differential Revision: D35530121

fbshipit-source-id: be06a23fc1266977b520e6b842ddee0e76f24908