Create the I2cDevIoTests and LassenQsfpI2cTests classes to contain unit tests for the I2cDevIo and LassenQsfpI2c classes
Summary:
The I2cDevIoTests and LassenQsfpI2cTests classes verify that the I2cDevIo and LassenQsfpI2c classes can handle errors properly (invalid I2C device files in this case because the unit tests will run on a devserver and not an actual Lassen switch). Read and write statistics are also tested for the LassenQsfpI2c class.

The MockI2cDevIo class will be created in the next diff to test valid read and write in the LassenQsfpI2c class.

Differential Revision: D35094347

fbshipit-source-id: c1b5b6937a33bbbf71b339ed372b0949b91116b7