Build a oper path-> publisher root abstraction
Summary:
Main idea is to centralize logic for publisher
root computation to a class and not scatter it around.
Today this logic is pretty simple so it didn't matter
much, however in the future (chassis) just assuming
first element of path as root won't do.

TODO: retire the convenience APIs in FsdbOperTreeMetadataTracker
which take in path. Have it deal with roots only for better
separation of responsibilities

Reviewed By: peygar

Differential Revision: D35632159

fbshipit-source-id: 883dfb13f422adecd91f28e5719c308dae259fb3