Use fastcli to gen tab complete script
Summary: I made some modifications to fastcli to support CLI11 and generate just a tab complete script. Using this to gen a script for fboss2. For now this does not support options, just our subcommands. If this proof of concept is good, I can extend to add our flags. next step is to install this on boxes/devservers

Reviewed By: shri-khare

Differential Revision: D34536965

fbshipit-source-id: aa0ec5d97f54d05d55a5f77e7ccc8ee0df9b66e2