add tab complete to rpm
Summary:
Adding tab complete script to rpm. Automatically adds a file to register the completions script to bash+zsh.
Had to throw the completion script into /usr/local/bin which is not ideal since it'll show up in user's PATH but not sure where else to put it to reference it from both completion.sh and completion.zsh

Reviewed By: shri-khare

Differential Revision: D35792926

fbshipit-source-id: 4650dc9e149ccf411e5a2cc41821cf228d486fd9