Run port prbs on right module id
Summary:
The common PRBS thrift interface takes the PortID as input. The port to module conversion gives 1 based transceiver id whereas the module objects are 0 based indexed so we need to correctly identify the module objects.

Update on 4/22/2022: The index change is taken care of client like wedge_qsfp_util CLI in later diff so that change removed from this diff

Differential Revision: D35808190

fbshipit-source-id: 115e33827f4f64cf6e2c4dcc2ac37533346d548f