Flush publisher queue on any state transition that's not connect
Summary:
When a stream disconnects or gets cancelled
its erroneous to still publisher queue. Post reconnect we
require a full sync. If we retain the items in publish
queue they will be sent to server before the full sync.
Server will take the first of these items as a full
sync and make the publisher root ready

Reviewed By: peygar

Differential Revision: D35852972

fbshipit-source-id: e7307bc1ef1de51e4e52aec93264dd03e30f81ed