Create the publishQueue only after connect, reset it on disconnect
Summary:
Publish queue now determines whether publisher is ready
to stream data to server or not. This way we don't run the
risk of accumulating data in publisher queue when the stream
is not connected.

TODO: Reset publisher queue on write errors as well. That way
on any write error publisher will fall back to reconnect
protocol

Reviewed By: peygar

Differential Revision: D35905875

fbshipit-source-id: bd361ea3a89cfb9ff3de2d857068bd8f79bbb53b