fix LogThriftCall for connections w/ cached cert
Summary:
I noticed some unexpected "unknown" identities in one of our
tools. After investigating a bit, it turns out that some transports
break the normal cert extraction from contexts, as the cert may be
cached at a lower layer. This includes the TLS1.3 cert implementation
(Fizz). The symptom here is that the first time we connect from a tool
to a service the library will see the cert and get a readable name,
but the subsequent times will just print "unknown".

This diff changes from just using getPeerCommonName on the connection
context to reusing some common auth libraries. Unfortunately, these
are not open source so I needed to split this library in to an
internal and oss version.

Reviewed By: jasmeetbagga

Differential Revision: D35904919

fbshipit-source-id: 9eb4341ea65288b9fd231c481f3470e153cb4d1c