On queue write error (overflow) reset the queue
Summary:
As titled. With a queue reset, we will break the publish
service loop and force the reconnect + full resync protocol.
This makes overflow handling be the same as reconnect, thus
minimizing the chances of bugs.

TODO: Since now we are resetting the queue, should we even
throw the error back on overflow? Service loop breakage will
already cause a disconnect + reconnect sequence.

Reviewed By: peygar

Differential Revision: D35916511

fbshipit-source-id: ba160912a78b846abe0fd2a0a9983826a8ea1ad4