Not consider it safe to reset if there's no programmed ports in cache
Summary:
https://fb.workplace.com/groups/1015730552263827/permalink/1340854456418100/
We noticed during this push, several ports flapping on Yamp.
For these ports, they all have smbus issue which causes
qsfp_service not able to detect the management interface type on a present transceivers.
But the current areAllPortsDown() doesn't consider the case that it's called before programming ports,
which is exactly what happened here in the push when it restarted qsfp_service.

Therefore, just like how we do if the cache doesn't have such transceiver, return false if there's no programmed ports in the cache

Differential Revision: D35993583

fbshipit-source-id: 7cfde4ab6575bb07619d1a7842c94252d4424ff0