lassen platform mapping with lane sort
Summary:
As titled, sort lanes to generate lassen platform mapping.

Port 29 and Port 31 are called quad lanes and we should use the bottom 4 lanes for port bring up.

- port 29 - lanes 12 through 15 for 40G, 100G and 200G
- port 31 - lanes 4 through 7 for 40G, 100G and 200G

NOTE: We did not observe the issue on port 29 because 29 is configured as OPTICS and we already sort the lanes to generate the lassen platform mapping for uplinks

Differential Revision: D36045795

fbshipit-source-id: 73d7ce763eff13e3d873a56e97a2e26b7ac6d149