Stop using EventBase in FbFpgaI2cController functions
Summary:
As we discussed in the group chat and S268898, the issue that caused such sev is due to the improper way of using EventBase in qsfp_service.
More details can be found in the root cause of S268898.

As for this diff, I decided to remove using the internal EventBase in these FbFpgaI2cController read/write functions as:
- They are already blocking and synchronized, so no point to throw them into the EventBase to introduce extra latency
- These functions are very low-level basic function as it actually the one to access the transceiver hardware, it might not be easy for the users to control the EventBase which could cause the chain issue like the sev.

Reviewed By: birdsoup

Differential Revision: D35981926

fbshipit-source-id: a8cd3b4f268b3dd4f402b3ef21452c93e5c79757