Added helper functions to read/write a SFF field
Summary:
This diff introduces helper functions for reading/writing from a SFF field (which includes the information about page, length, offset). The helper functions take care of setting the page also. The user can choose to skip changing page if the user is doing a batch operation and knows the page will not change in the middle.

The aim is to avoid cases where we accidentally skip changing the page sometimes. In the next diffs, I'll also add a CmisModule version of this helper function. In CmisModule, we need to skip changing page altogether in case of a miniphoton2 module.

Differential Revision: D35902786

fbshipit-source-id: c14de131f7a82300f9f53aa93eba073c1762b6f7