Introduce option to get egressPoolSharedLimit Native BCM
Summary:
For ECN test to validate that no drops happen when traffic is in
queue, we need to figure how much can a queue build. This is dependent on the
service pool shared buffer limit which is programmed by SDK based on sum of all
mins used in the system and hence including an API to read this information
from SDK for use in test.

Differential Revision: D35806106

fbshipit-source-id: 57d0626a9e0fe6557159a5864fde5e0799871a6c