Set default FEC counter to 0 rather than -1
Summary: Currently we default FEC counters to -1 and accumulate from there. This is misleading since a UCW count of -1 looks like FEC just isn't supported, when it really means that we haven't seen an error. Similarly a UCW count of 0 looks like there have been no errors, but currently means that we've seen 1 UCW. To fix this, if we successfully read the fec counters from the chip, we should set the fec counter to 0 before accumulating.

Differential Revision: D35982022

fbshipit-source-id: a63f836e23d709567f90a03dfccd1f1ef66ccb64