Fetch TransceiverInfo from qsfp_service instead of cache
Summary:
Sometimes we see syncPorts time out if there's enough refreshing/port programming that occurs during it. When this happens in our test setup, we fail to fetch TransceiverInfo for the qsfpAche and will then not have any ports to run the test on, which sometimes causes failures. This diff will retry a few times to fetch TransceiverInfo before running the test so we don't run into this issue.

Context: https://fb.workplace.com/groups/1015730552263827/permalink/1339378076565738/

Reviewed By: rajank7, joseph5wu

Differential Revision: D35946227

fbshipit-source-id: 8aa52d72dfffee3991c79146022c9c6b8db9df99