Give specific error message instead of "condition was never satisfied"
Summary:
Current message:
```
I0504 23:34:33.976765 1816324 LinkTest.cpp:248]  No lldp neighbors on : eth1/15/1
...
C++ exception with description "Verify with retry failed, condition was never satisfied" thrown in the test body.
```
Wasn't able to reproduce problem on same setup. So ignoring the Conveyer failure and just improve the error for ease of debugging in the future.

Differential Revision: D36206875

fbshipit-source-id: 4b26ca0d3f4c22e639d85b55c23d23c57e8a9dd2