Use correct "Application Select" value 1 to 15 in code
Summary:
As seen in task T120725078, the test fails with the following message.

  F0516 03:37:44.986644 60750 CmisModule.cpp:922] Check failed: channel <= 8 (9 vs. 8)

I think this happens because in CMisModule::numMediaLanes() it does not determine that the optics has only 4 lanes.

The optical current “application select" value is read from 01.206. The valid values are 1-15,  the code indexes ApSel-1 to registers 0.86 to fnd correct “media application code”. I think in case of unit tests we return 0 so it indexes -1 causing the issue.

Besides our code should not recognize “application select" value 0 as valid value because 0 means the current application is not selected yet. So I made the code changes here to treat "application select" value 0 as not valid.

Differential Revision: D36536569

fbshipit-source-id: 085451b4c89e08da8c665072ed3b5f13cbb1bfb7