Rename switch state Transceiver to TransceiverSpec
Summary:
As titled. This is towards publish TransceiverSpec deltas to fsdb.

To achieve that, we'll need to link the `Transceiver.h` header file to `StateDelta.h`. Unfortunately it is also linked with qsfp service, which also has a definition of Transceiver class.

Therefore, renaming Transceiver class in agent state to be TransceiverSpec class. This should avoid the collision of both of these two classes.

Reviewed By: peygar

Differential Revision: D36644751

fbshipit-source-id: 904909a082cdfab629f419fdf171a681edd57f9b