Use WITH_RETRIES_N_TIMED in PrbsTest and also extend the timeouts
Summary:
WITH_RETRIES_N_TIMED is preferred over checkWithRetry since it provides better logs on failures.

Also increased the timeout from default 10seconds to 60seconds to allow qsfp_service to successfully refresh a transceiver. Previously, the test would fail occasionally due to failed refreshes because of i2c failures. Because the refresh on cmis modules is slow and happens ~ 15seconds, we never get a successful refresh for the duration of the test and that causes test failures.

Reviewed By: rajank7

Differential Revision: D36781806

fbshipit-source-id: 67a62ed77f0287696ef2da5587c45b70901237dc