Delete Accton xphy code
Summary:
These commands are already broken by moving the xphy logic to the qsfp service (the xphy commands are in the wedge_agent, but the agent no longer can access the xphy objects). Since we're not using these anyways, lets remove them

We also need to remove or completely refactor these if we want to remove the asic sdk from the qsfp_service.

Reviewed By: rajank7

Differential Revision: D36679122

fbshipit-source-id: 38354810089adb9e497e40b900ee755ca79047b3