Pass wedgeManager to all direct i2c functions
Summary: We currently fail the nullptr CHECK in Transceiver.h when using --direct_i2c because we don't pass the wedgeManager in. Since we already create wedgeManager in qsfp_util, we can just pass it along with the bus pointer so we can use it when creating the QsfpModule object.

Differential Revision: D36791363

fbshipit-source-id: ed09b20c18f8acfb834acb602905d90f8adb9f92