Add interface id as optional parameter in redirect nexthop
Summary:
Adding an optional interface id in redirect nexthop thrift

In case of local link drain, there will not be any event that agent can act on to deactivate the acl if we use the neighbor's interface address. To handle this, we need the redirection nexthop ip address also to be advertised by BGP so that it will be withdrawn during drain. The idea is to announce an anycast ip through bgp from all CTSWs. When programming ACL, the redirection config will include the outgoing interface along with the bgp advertised  nexthop ip. After fib lookup of the bgp advertised nexthop, agent will pick the outgoing path from nexthop set based on the interface id. This allows us to track local link drain also.

Reviewed By: jasmeetbagga

Differential Revision: D36598817

fbshipit-source-id: 2f529957766cc07ddd7ca004b8327c5262bd6770