Add new test to validate ECN traffic is not dropped in queue
Summary:
New test to make sure that when a queue is built up with ECN packets,
no drops happen in the queue except for when the queue limit is hit. For EBRO,
ensure that queue can build up to the limit and for Broadcom, make sure that
queue can build to 99.9% of the dynamic alpha based threshold for service pool
shared buffer limit.

Reviewed By: zechengh09

Differential Revision: D35806104

fbshipit-source-id: 713231c9034dd0b7e4d21739848c832dc93ba696