rackmon2: Fix large timeouts observed on wedge400
Summary:
The current write method goes in 4 steps.
1. disableRead
2. write request
3. wait for LSR assert.
4. enableRead
5. Read response
Normally, a response would arrive after #4 thus,
marking the transaction as successful.
But occationally (1/10 times) the response
arrives before #4 thus causing the device to
completely miss the response -- resulting in a timeout.

This change, replaces this with
1. Keeping read enabled throughout.
2. ignore LSR changes (Do not do #3).
This change completely eliminates timeouts on wedge400+ with
an FTDI USB device which do not need this sequencing.

But we would still need this functionality for wedge100 where we use
the native aspeed device (Pending a verification on wedge100. If it
is not needed even on that platform, that functionality can be
completely stripped.)

Imported from: D35445732

Reviewed By: jasmeetbagga

Differential Revision: D36388930

fbshipit-source-id: 3bfe0aa1e633732fecd48a14b478da01dc0a8f8b