Add the commands in kAdditionalcommandTree to the reverse help tree. Also finished implementation of the help command handler.
Summary:
Added the commands from the kAdditionalCommandTree to the reverse help tree. Also tested the correctness code on a few dummy inputs.

This is a first draft of a complete implementation of the help command!

Command: fboss2 help <objectName>
This will produce the multi-level help tree for all verbs and subcommands that can be executed on the requested objectName.

Example: fboss2 help interface

Reviewed By: peygar

Differential Revision: D36814206

fbshipit-source-id: 0a49ea04d8149dda4d72ac3c64e65d1e28345b6f